+++ light-std/src/lstd/graphics/buffer.cpp
@@ -4,22 +4,22 @@
 
 LSTD_BEGIN_NAMESPACE
 
-void buffer_layout::add(string name, gtype type, size_t count, bool normalized) {
-    size_t sizeInBits = get_size_of_base_gtype_in_bits(type);
+void buffer_layout::add(string name, gtype type, s64 count, bool normalized) {
+    s64 sizeInBits = get_size_of_base_gtype_in_bits(type);
 
     count *= get_count_of_gtype(type);
 
     assert(TotalSize <= numeric_info<u32>::max());
-    Elements.append({name, get_scalar_gtype(type), sizeInBits, normalized, (u32) count, (u32) TotalSize});
+    Elements.append({name, get_scalar_gtype(type), sizeInBits, normalized, count, (u32) TotalSize});
 
     if (sizeInBits == 1) sizeInBits = 8;
     TotalSize += (sizeInBits / 8) * count;
 }
 
-void buffer_layout::add_padding(size_t bytes) { TotalSize += bytes; }
+void buffer_layout::add_padding(s64 bytes) { TotalSize += bytes; }
 
 extern buffer::impl g_D3DBufferImpl;  // Defined in d3d_buffer.cpp
-void buffer::init(graphics *g, buffer_type type, buffer_usage usage, size_t size, const char *data) {
+void buffer::init(graphics *g, buffer_type type, buffer_usage usage, s64 size, const char *data) {
     Graphics = g;
     Type = type;
     Usage = usage;
