+++ light-std/src/lstd/math/no_init.h
@@ -8,13 +8,13 @@ struct no_init_t {};
 constexpr no_init_t no_init = {};
 
 namespace impl {
-template <typename MathT, size_t... Is>
+template <typename MathT, s64... Is>
 stack_array<MathT, sizeof...(Is)> make_stack_array_helper(index_sequence<Is...>) {
     return {((void) Is, MathT(no_init))...};
 }
 }  // namespace impl
 
-template <typename MathT, size_t N>
+template <typename MathT, s64 N>
 stack_array<MathT, N> make_stack_array_of_uninitialized_math_type() {
     return impl::make_stack_array_helper<MathT>(make_index_sequence<N>());
 }
